move formatExportableSessions function in /lib/network-exporters/formatters
